% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat-package.R
\docType{data}
\encoding{UTF-8}
\name{distdata}
\alias{distdata}
\alias{dstrokedit}
\alias{dyehli}
\title{Precomputed kanji distances}
\format{
Sparse matrices containing in each row the distances between a key kanji and its ten nearest neighbors.\if{html}{\out{<br>}}
For \code{dstrokedit}, these are the stroke edit distances according to Yencken and Baldwin (2008).\if{html}{\out{<br>}}
For \code{dyehli}, these are the bag-of-radicals distances according to Yeh and Li (2002).

All pre-2010 jouyou kanji that are also post-2010
jouyou kanji are included. The indices are those from \code{\link{kbase}}.
}
\source{
Datasets from \url{https://lars.yencken.org/datasets}, made available under the
Creative Commons Attribution 3.0 Unported licence.

Computed as part of \emph{Yencken, Lars (2010)
\href{https://lars.yencken.org/papers/phd-thesis.pdf}{Orthographic support for passing the reading hurdle in Japanese}.
PhD Thesis, University of Melbourne, Melbourne, Australia}.
}
\usage{
dstrokedit

dyehli
}
\description{
Precomputed kanji distances
}
\examples{
# Look up characters with smallest stroke edit distance to 部.
bu_index <- match("部", kbase$kanji)
non_zero <- which(dstrokedit[bu_index,] != 0)
rbind(kbase[non_zero,]$kanji, dstrokedit[non_zero,bu_index])

# Look up characters with smallest bag-of-radicals distance to 部.
bu_index <- match("部", kbase$kanji)
non_zero <- which(dyehli[bu_index,] != 0)
rbind(kbase[non_zero,]$kanji, dyehli[non_zero,bu_index])

}
\references{
Yeh, Su-Ling
and Li, Jing-Ling (2002). Role of structure and component in judgements of
visual similarity of Chinese characters. \emph{Journal of Experimental Psychology:
Human Perception and Performance}, \strong{28}(4), 933–947.

Yencken, Lars, & Baldwin, Timothy (2008). Measuring and predicting orthographic associations:
Modelling the similarity of Japanese kanji. In: \emph{Proceedings of the 22nd International Conference on Computational
Linguistics (Coling 2008)}, pp. 1041-1048.
}
\keyword{datasets}
