% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat-package.R
\docType{data}
\name{kreadmean}
\alias{kreadmean}
\title{Kanji readings and meanings}
\format{
An object of class \code{list} of length 13108.
}
\source{
KANJIDIC2 file by Jim Breen and The Electronic Dictionary Research
and Development Group (EDRDG)\cr
\url{http://www.edrdg.org/wiki/index.php/KANJIDIC_Project}\cr
The use of this data is covered by the Creative Commons BY-SA 4.0 License.
}
\usage{
kreadmean
}
\description{
Data set of all kanji readings and meanings from the KANJIDIC2 dataset in an R list
format. For convenient access to this data use function \code{\link{lookup}}.
}
\keyword{datasets}
