% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanji1.R
\name{plotkanji}
\alias{plotkanji}
\title{Plot a kanji}
\usage{
plotkanji(
  kanji,
  device = "default",
  family = NULL,
  factor = 10,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{kanji}{a single character, string or numeric specifying a single kanji.}

\item{device}{the type of graphics device where the kanji is plotted. Defaults to the
user's default type according to \code{getOption("device")}. Can be set to
"current", in which case the kanji is plotted on the currently active device.}

\item{family}{the font family used for writing the kanji. Make sure to add the font
first by using \code{\link[sysfonts]{font_add}}. See details.}

\item{factor}{a maginification factor applied to the font size (typically 12 points).}

\item{width, height}{the dimensions of the device.}

\item{...}{further parameters passed to the function opening the device (such as a
file name for devices that create a file).}
}
\description{
Plot a kanji
}
\details{
This function allows to "draw" a kanji on a number of possible graphics devices
in an arbitrary font that is installed on the user's system and has been added
to the database in package \code{sysfonts}. For the latter \code{\link[sysfonts]{font_add}} or
\code{\link[sysfonts]{font_families}} to verify what fonts are available. Locating user installed
fonts varies from system to system and finding the path needed for \code{\link[sysfonts]{font_add}}
can sometimes be tricky. The suggested package \code{systemfonts} (NOT the same as \code{sysfonts})
offers the helpful command \code{\link[systemfonts]{match_font}}, which provides the font path based
on the name of the font family (and if desired the font face).

\if{html}{\out{<div class="sourceCode">}}\preformatted{     The function uses then the package \code{showtext} to write the kanji in a large font at the
     center of a new device of the specified type. If no font family is provided, the CJK font 
     WenQuanYi Micro Hei that that comes with the package showtext is used. This font targets
     Chinese writing and some strokes will therefore look odd for Japanese characters. We strongly
     advised that a Japanese font is used.
     
     For further details on CJK characters and computers, see the corresponding vignette (TO DO). 
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
plotkanji("滝")   # uses Chinese font!}

}
