% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{samplekan}
\alias{samplekan}
\title{Sample kanji from a set}
\usage{
samplekan(
  set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"),
  size = 1,
  replace = FALSE,
  prob = NULL
)
}
\arguments{
\item{set}{a character string specifying the set of kanjis to sample from.}

\item{size}{a positive number, the number of samples.}

\item{replace}{logical. Sample with replacement?}

\item{prob}{currently without effect.}
}
\value{
a vector of length \code{size} containing the individual characters
}
\description{
Sample kanji from a set
}
\examples{
(sam <- samplekan(size = 10))
lookup(sam)
}
