% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanjistat-package.R
\docType{data}
\encoding{UTF-8}
\name{yehLiDistance}
\alias{yehLiDistance}
\title{Bag-of-Radicals Distance according to Yeh and Li}
\format{
\code{yehLiDistance} is a sparse matrix containing
the bag-of-radicals distance according to Yeh and Li (2002) for
its ten nearest neighbors. All pre-2010 jouyou kanji that are also post-2010
jouyou kanji are included. The indices are those from \code{\link{kbase}}.
}
\source{
Dataset from \url{https://lars.yencken.org/datasets} under
Creative Commons Attribution 3.0 Unported, implementing "Yeh, Su-Ling
and Li, Jing-Ling. 2002. Role of structure and component in judgements of
visual similarity of Chinese characters. Journal of Experimental Psychology:
Human Perception and Performance, 28(4):933–947".
}
\usage{
yehLiDistance
}
\description{
Bag-of-Radicals Distance according to Yeh and Li
}
\examples{
# Look up characters with smallest bag-of-radicals distance to 部.
bu_index <- match("部", kanjistat::kbase$kanji)
non_zero <- which(yehLiDistance[bu_index,] != 0)
rbind(kbase[non_zero,]$kanji, yehLiDistance[non_zero,bu_index])

}
\keyword{datasets}
